refactor (report): move report controller to folder report and fix validation
