+++ app/Http/Controllers/Report/StockReportController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Dashboard;
+namespace App\Http\Controllers\Report;
 
 use App\Exports\StockExportExcel;
 use App\Http\Controllers\Controller;
